% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dlib.R
\docType{package}
\name{dlib}
\alias{dlib}
\alias{dlib-package}
\title{R integration with dlib}
\description{
This package makes the dlib C++ library - which contains both generic tools for
             software development, and BSL-licensed machine learning functionality - accessible
             from R, so that it can be conveniently included in other packages.
}

