% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{inlineCxxPlugin}
\alias{inlineCxxPlugin}
\title{Rcpp plugin for dlib}
\usage{
inlineCxxPlugin()
}
\value{
a list
}
\description{
Rcpp plugin for dlib
}
\examples{
library(Rcpp)
library(inline)
library(dlib)
getPlugin("dlib")
}
