% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{surf_descriptor}
\alias{surf_descriptor}
\title{Example of using dlib to extract SURF points}
\usage{
surf_descriptor(file_name)
}
\arguments{
\item{file_name}{path to a bmp image}
}
\value{
a list with the number of surf points found in the image
}
\description{
Example of using dlib to extract SURF points
}
